% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run-spagedi.R
\name{runSpagedi}
\alias{runSpagedi}
\title{Execute SPAGeDi through R console.}
\usage{
runSpagedi(input_name, output_name, perm = 999)
}
\arguments{
\item{input_name}{\(string\) - Name of SPAGeDi formatted input data file. Must be in
working directory.}

\item{output_name}{\(string\) - Name of output file. Will overwrite if file
already exists, or create new file if it doesn't}

\item{perm}{Number of permutations. Default is 999}
}
\value{
Does not return anything directly, but writes results of analysis to
 a tab-delimited text file at \code{output_name}.
}
\description{
Calls SPAGeDi executable via system() function
}
\details{
SPAGeDi must be installed on your system, and able to be run from the
 Shell/Terminal. Function is currently configured to run a basic spatial
 autocorrelation analysis between individuals using the kinship coefficient
 of Loiselle et al. 1995, testing for significance, and jacknifing over loci.
}

